+++ akshare/__init__.py
@@ -1904,9 +1904,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.4.81 add: add fund_portfolio_change_em interface
 1.4.82 fix: fix stock_a_code_to_symbol interface
 1.4.83 add: add stock_us_code_table_fu interface
+1.4.84 fix: fix fund_portfolio_change_em interface
 """
 
-__version__ = "1.4.83"
+__version__ = "1.4.84"
 __author__ = "AKFamily"
 
 import sys
@@ -2630,7 +2631,7 @@ from akshare.futures.futures_to_spot import (
 """
 基金持仓
 """
-from akshare.fund.fund_em_portfolio import fund_portfolio_hold_em, fund_portfolio_change_em
+from akshare.fund.fund_portfolio_em import fund_portfolio_hold_em, fund_portfolio_change_em
 
 """
 债券概览