+++ akshare/fund/fund_portfolio_em.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2022/3/20 20:54
-Desc: 天天基金网-基金档案-投资组合-基金持仓
+Date: 2022/3/21 17:40
+Desc: 天天基金网-基金档案-投资组合
 http://fundf10.eastmoney.com/ccmx_000001.html
 """
 import pandas as pd
@@ -60,29 +60,37 @@ def fund_portfolio_hold_em(symbol: str = "162411", date: str = "2020") -> pd.Dat
             ]
         ]
         big_df = big_df.append(temp_df, ignore_index=True)
-    big_df['占净值比例'] = pd.to_numeric(big_df['占净值比例'], errors="coerce")
-    big_df['持股数'] = pd.to_numeric(big_df['持股数'], errors="coerce")
-    big_df['持仓市值'] = pd.to_numeric(big_df['持仓市值'], errors="coerce")
-    big_df['序号'] = range(1, len(big_df)+1)
+    big_df["占净值比例"] = pd.to_numeric(big_df["占净值比例"], errors="coerce")
+    big_df["持股数"] = pd.to_numeric(big_df["持股数"], errors="coerce")
+    big_df["持仓市值"] = pd.to_numeric(big_df["持仓市值"], errors="coerce")
+    big_df["序号"] = range(1, len(big_df) + 1)
     return big_df
 
 
-def fund_portfolio_change_em(symbol: str = "003567", date: str = "2020") -> pd.DataFrame:
+def fund_portfolio_change_em(
+    symbol: str = "003567", indicator: str = "累计买入", date: str = "2020"
+) -> pd.DataFrame:
     """
     天天基金网-基金档案-投资组合-重大变动
     http://fundf10.eastmoney.com/ccbd_000001.html
     :param symbol: 基金代码
     :type symbol: str
+    :param indicator: choice of {"累计买入", "累计卖出"}
+    :type indicator: str
     :param date: 查询年份
     :type date: str
     :return: 重大变动
     :rtype: pandas.DataFrame
     """
+    indicator_map = {
+        "累计买入": "1",
+        "累计卖出": "2",
+    }
     url = "http://fundf10.eastmoney.com/FundArchivesDatas.aspx"
     params = {
         "type": "zdbd",
         "code": symbol,
-        "zdbd": "1",
+        "zdbd": indicator_map[indicator],
         "year": date,
         "rt": "0.913877030254846",
     }
@@ -98,7 +106,9 @@ def fund_portfolio_change_em(symbol: str = "003567", date: str = "2020") -> pd.D
     for item in range(len(item_label)):
         temp_df = pd.read_html(data_json["content"], converters={"股票代码": str})[item]
         del temp_df["相关资讯"]
-        temp_df["占期初基金资产净值比例（%）"] = temp_df["占期初基金资产净值比例（%）"].str.split("%", expand=True).iloc[:, 0]
+        temp_df["占期初基金资产净值比例（%）"] = (
+            temp_df["占期初基金资产净值比例（%）"].str.split("%", expand=True).iloc[:, 0]
+        )
         temp_df["季度"] = item_label[item]
         temp_df.columns = [
             "序号",
@@ -119,13 +129,13 @@ def fund_portfolio_change_em(symbol: str = "003567", date: str = "2020") -> pd.D
             ]
         ]
         big_df = big_df.append(temp_df, ignore_index=True)
-    del big_df['序号']
+    del big_df["序号"]
     big_df.reset_index(inplace=True)
-    big_df['index'] = big_df.index + 1
+    big_df["index"] = big_df.index + 1
     big_df.rename(columns={"index": "序号"}, inplace=True)
 
-    big_df['本期累计买入金额'] = pd.to_numeric(big_df['本期累计买入金额'], errors="coerce")
-    big_df['占期初基金资产净值比例'] = pd.to_numeric(big_df['占期初基金资产净值比例'], errors="coerce")
+    big_df["本期累计买入金额"] = pd.to_numeric(big_df["本期累计买入金额"], errors="coerce")
+    big_df["占期初基金资产净值比例"] = pd.to_numeric(big_df["占期初基金资产净值比例"], errors="coerce")
     return big_df
 
 
@@ -133,5 +143,12 @@ if __name__ == "__main__":
     fund_portfolio_hold_em_df = fund_portfolio_hold_em(symbol="162411", date="2020")
     print(fund_portfolio_hold_em_df)
 
-    fund_portfolio_change_em_df = fund_portfolio_change_em(symbol="003567", date="2020")
+    fund_portfolio_change_em_df = fund_portfolio_change_em(
+        symbol="003567", indicator="累计买入", date="2020"
+    )
+    print(fund_portfolio_change_em_df)
+
+    fund_portfolio_change_em_df = fund_portfolio_change_em(
+        symbol="003567", indicator="累计卖出", date="2020"
+    )
     print(fund_portfolio_change_em_df)