+++ akshare/stock_feature/stock_us_hist_futunn.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2021/12/2 17:26
+Date: 2022/3/21 15:26
 Desc: 富途牛牛-行情-美股-每日行情
 https://www.futunn.com/stock/HON-US?seo_redirect=1&channel=1244&subchannel=2&from=BaiduAladdin&utm_source=alading_user&utm_medium=website_growth
 """